+++ apps/ungdomsytelse-deltaker/src/apps/søknad/hooks/api/useSendSøknad.ts
@@ -1,9 +1,8 @@
 import { ungdomsytelse } from '@navikt/k9-brukerdialog-prosessering-api';
 import { ApiError } from '@navikt/ung-common';
+import { commonQueries } from '@shared/api/queries/commonQueries';
+import { sendSøknad } from '@søknad/api/søknad/sendSøknad';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-
-import { commonQueries } from '../../../../api/queries/commonQueries';
-import { sendSøknad } from '../../api/søknad/sendSøknad';
 
 export const useSendSøknad = () => {
     const queryClient = useQueryClient();