+++ apps/ungdomsytelse-deltaker/src/apps/søknad/pages/VelkommenPage.tsx
@@ -1,10 +1,10 @@
 import { ArrowRightIcon } from '@navikt/aksel-icons';
 import { BodyLong, Box, Button, Checkbox, CheckboxGroup, VStack } from '@navikt/ds-react';
+import { AppText, useAppIntl } from '@shared/i18n';
+import DefaultPageLayout from '@shared/pages/layout/DefaultPageLayout';
+import getLenker from '@shared/utils/lenker';
 import { useState } from 'react';
 
-import { AppText, useAppIntl } from '../../../i18n';
-import getLenker from '../../../utils/lenker';
-import DefaultPageLayout from '../../innsyn/pages/layout/DefaultPageLayout';
 import ExternalLink from '../components/external-link/ExternalLink';
 import SøknadHeader from '../components/søknad-header/SøknadHeader';
 import VelkommenMelding from '../components/VelkommenMelding';