+++ apps/ungdomsytelse-deltaker/storybook/decorators/withDeltakerContext.tsx
@@ -1,12 +1,12 @@
 import { søkerResponseSchema } from '@navikt/sif-common-api';
 
-import { getScenarioMockData } from '../../mock/scenarios/scenarioMap';
+import { getScenarioMockData } from '../../mock/scenarios/scenarioer';
 import { ScenarioType } from '../../mock/scenarios/types';
 import { DeltakerContextProvider } from '../../src/context/DeltakerContext';
 import { deltakelsePeriodeSchema } from '../../src/types/DeltakelsePeriode';
 
 export const withDeltakerContext = (Story) => {
-    const { søker, deltakelser } = getScenarioMockData(ScenarioType.harSøkt);
+    const { søker, deltakelser } = getScenarioMockData(ScenarioType.søknadSendt);
     return (
         <DeltakerContextProvider
             søker={søkerResponseSchema.parse(søker)}