+++ apps/ungdomsytelse-deltaker/storybook/decorators/withSøknadContext.tsx
@@ -1,7 +1,7 @@
 import { Theme } from '@navikt/ds-react';
 import { barnResponseSchema } from '@navikt/sif-common-api';
 
-import { getScenarioMockData } from '../../mock/scenarios/scenarioMap';
+import { getScenarioMockData } from '../../mock/scenarios/scenarioer';
 import { ScenarioType } from '../../mock/scenarios/types';
 import { SøknadProvider } from '../../src/apps/søknad/context/SøknadContext';
 import { HarKontonummerEnum } from '../../src/apps/søknad/steg/oppsummering/oppsummeringUtils';
@@ -9,7 +9,7 @@ import { SøknadContextType } from '../../src/apps/søknad/types';
 import { DeltakelsePeriode, deltakelsePeriodeSchema } from '../../src/types/DeltakelsePeriode';
 import { SøkYtelseOppgave } from '../../src/types/Oppgave';
 
-const data = getScenarioMockData(ScenarioType.harIkkeSøkt);
+const data = getScenarioMockData(ScenarioType.søknad);
 
 export const withSøknadContext = (Story: any, context?: Partial<SøknadContextType>) => {
     const deltakelse = deltakelsePeriodeSchema.parse(data.deltakelser[0]);