+++ packages/k9-sak-innsyn-k9-sak-api/scripts/fix-generated-regex.mjs
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { fixAndFormatGeneratedCode } from '../../../codegenUtils.js';
+
+fixAndFormatGeneratedCode('src');