+++ packages/ung-deltakelse-opplyser-api-deltaker/src/index.ts
@@ -1,7 +1,7 @@
 export * from './deltaker/client';
-export * from './deltaker/client/zod.gen';
-export { zDeltakelseKomposittDto } from './deltaker/client/zod.gen';
+export * from './deltaker/client/client.gen';
 export { OppgaveStatus } from './deltaker/client/types.gen';
 export { Oppgavetype } from './deltaker/client/types.gen';
-export * from './deltaker/client/client.gen';
+export * from './deltaker/client/zod.gen';
+export { zDeltakelseKomposittDto } from './deltaker/client/zod.gen';
 export * from './ungDeltakelseOpplyserApiClient';