+++ packages/appstatus-react-ds/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @navikt/appstatus-react-ds
 
+## 2.20.49
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 2.20.48
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+    - @navikt/sif-common-hooks@0.3.6
+
+## 2.20.47
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+    - @navikt/sif-common-hooks@0.3.6
+
 ## 2.20.46
 
 ### Patch Changes