+++ packages/sif-common-amplitude/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-amplitude
 
+## 2.28.40
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 2.28.39
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 2.28.38
 
 ### Patch Changes