+++ packages/sif-common-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-utils
 
+## 3.49.38
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 3.49.37
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 3.49.36
 
 ### Patch Changes