+++ packages/ung-deltakelse-opplyser-api-deltaker/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @navikt/ung-deltakelse-opplyser-api
 
+## 0.1.32
+
+### Patch Changes
+
+- @navikt/sif-common-api@0.4.58
+
+## 0.1.31
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/sif-common-api@0.4.57
+
+## 0.1.30
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/sif-common-api@0.4.56
+
 ## 0.1.29
 
 ### Patch Changes